One update of services max per dispatching.
Now, the useful informations (state and newest record) are locally stored
in a hstore, always contening the values corresponding to the max timestamp.
At the end of dispatch_record(), we update every services found in this hstore,
if the local max timestamp is more than the one in the wh_nagios.services table.