Handle perfdata file with empty perfdata.
Some recent version of Nagios can generates perfdata file with empty perfdata.
Handle this case to allow reporting a status change if any to the UI.

A special label named " " is used for that purpose, either in nagios_dispatcher
and dispatch_record function.

Thanks to Kyungho Kim for the report and help providing such perfdata files.