Don't truncate metric table on duplicate key.
Previously, the wh_nagios.create_partition_on_insert_metric trigger would
truncate the target table if it already exists.  The only reason the target
table would exists (except someone creating such a table) is during a restore.
In this case, truncating the target table is a bad idea, because it would
prevent from restoring the data. So just ignore the error and continue.