Change the perfdata storage.
Previously, each metric was stored in a dedicated partition.  This
design was bad and lead to multiple issue with the growing number of
tables and indexes.  Instead, only store one partition per service.
While not being ideal, it should still greatly lower the number of
objects in the OPM database.